backend/requirements.txt
mss>=9.0.1
opencv-python>=4.8.0
numpy>=1.24.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
websockets>=12.0
pillow>=10.0.0
